+++ src/app/api/contact/route.ts
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 import { NextRequest, NextResponse } from 'next/server';
 
-import connectDb from '@/app/lib/dbConnect';
+import connectDb from '@/lib/dbConnect';
 import Contact from '@/models/Contact';
 
 export async function POST(req: NextRequest) {
